- add player to scene.

+++ Snake3D/Snake3D/Common/AppConsts.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct AppConsts {
     public static let MAP_SIZE =                                        12
-    public static let MOVEMENT_DELAY: Float =                           0.5
+    public static let MOVEMENT_DELAY: Double =                          0.5
     public static let SCALE_FACTOR =                                    10
     public static let SCALE_FACTOR_SEGMENT: Float =                     0.85
 }
